Recuperacao da 1a versao do Produto Java

+++ PastaFontesApp/Produto.java
@@ -1,34 +1,34 @@
-// Classe de objetos que guarda daods de produtos e servicos oferecidos no estacionamento
-// Exemplos de produtos e servicos: estacionar veiculo; lavar veiculo; trocar oleo
 
 public class Produto {
 
-	private int peso;
+	private double peso;
 	private double altura;
-
+	
 	public Produto() {
-
+		
 	}
-
-	public Produto(int peso, double altura) {
-		this.peso = peso;
-		this.altura = altura;
+	
+	public Produto(double peso, double altura) {
+		this.peso=peso;
+		this.altura=altura;
 	}
-
-	public int getPeso() {
+	
+	public double getPeso() {
 		return peso;
 	}
-
-	public void setPeso(int peso) {
+	
+	public void setPeso(double peso) {
 		this.peso = peso;
 	}
-
+	
 	public double getAltura() {
 		return altura;
 	}
-
+	
 	public void setAltura(double altura) {
 		this.altura = altura;
 	}
-
+	
+	
+	
 }
